+++ src/vpe/mapping.py
@@ -306,15 +306,59 @@ def imap(
         pass_info=pass_info, vim_exprs=vim_exprs)
 
 
-# TODO: Rename this and other handlers ...Mixin?
 class KeyHandler:
-    """Mix-in to support mapping key sequences to methods."""
+    """Mix-in to support mapping key sequences to methods.
 
+    This can be used as a simple base class, but also as a mix-in. For
+    example:<py>:
+
+        class MyScratchBuffer(ScratchBuffer, KeyHandler):
+            def __init__(*args, **kwargs):
+                super().__init__(*args, **kwargs)
+                self.auto_map_keys()
+                ...
+
+    The `mapped` static method is used to decorate the methods that you wish to
+    handle mappings. Some examples:<py>:
+
+        @KeyHandler.mapped('<C-F1>')
+        def show_general_help(self) -> None:
+            ...
+
+        @KeyHandler.mapped('<Leader>h')
+        def show_help_for_word_under_cursor(self) -> None:
+            ...
+
+        # This mapping is forced to be global by the ``buffer=False`` argument.
+        @KeyHandler.mapped('<F12>', buffer=False)
+        def list_buffers(self) -> None:
+            ...
+
+    The `auto_map_keys` method must be called to set up the mappings, typically
+    within the class ``__init__`` method. If the main class is a subclass of
+    `Buffer` (such as MyScratchBuffer above) then the mappings are set up just
+    for that buffer, unless ``buffer`` argument was specified for the `mapped`
+    method. Otherwise the mappings default to global.
+    """
     #: A list of key sequences, mapping mode and docstring.
     map_info: Final[list[tuple[str, str, str]]]
 
-    def auto_map_keys(self, *, pass_info: bool = False):
-        """Set up mappings for methods."""
+    def auto_map_keys(self, *, pass_info: bool = False) -> None:
+        """Set up mappings for methods.
+
+        :pass_info:
+            If set then each key handler method will be invoked with a
+            `MappingInfo` object. This is useful if you want a single method to
+            handle several mappings, but behave differently depening on which
+            mapping was triggered.
+        """
+        if isinstance(self, Buffer):
+            with core.temp_active_buffer(self):
+                self._do_auto_map_keys(pass_info=pass_info, buffer=True)
+        else:
+            self._do_auto_map_keys(pass_info=pass_info, buffer=False)
+
+    def _do_auto_map_keys(self, *, pass_info: bool, buffer: bool) -> None:
         def is_method(obj):
             return inspect.ismethod(obj) or inspect.isfunction(obj)
 
@@ -333,7 +377,10 @@ class KeyHandler:
                 info = getattr(method, '_keymappings_', None)
                 if info is not None:
                     for mode, keyseq, docstring, kwargs in info:
-                        kmap(mode, keyseq, method, **kwargs)
+                        kw = kwargs.copy()
+                        if 'buffer' not in kw:
+                            kw['buffer'] = buffer
+                        kmap(mode, keyseq, method, **kw)
                         map_info.append((keyseq, mode, docstring))
 
     @staticmethod